automation/bookstack-updater/requirements.txt
requests>=2.31.0
pyyaml>=6.0
jinja2>=3.1.0
discord-webhook>=1.3.0
